First renderer with skia, that actually works.

+++ Cable.App/Models/Data/Nodes/Geometry/RectangleNode.cs
@@ -14,7 +14,7 @@ public partial class RectangleNode : NodeData<Geometry2D>
     public override Geometry2D GetTypedOutput()
     {
         Vector2[] vertices = [new(0, 0), new(_width, 0), new(_width, _height), new(0, _height)];
-        int[] indices = { 0, 1, 2, 0, 2, 3 };
+        ushort[] indices = { 0, 1, 2, 0, 2, 3 };
 
         return new Geometry2D(vertices, indices);
     }
