+++ Cable.Data/Types/Geometry2D.cs
@@ -2,8 +2,8 @@
 
 namespace Cable.Data.Types;
 
-public readonly struct Geometry2D(Vector2[] vertices, int[] indices) : ICableDataType
+public readonly struct Geometry2D(Vector2[] vertices, ushort[] indices) : ICableDataType
 {
     public readonly Vector2[] Vertices = vertices;
-    public readonly int[] Indices = indices;
+    public readonly ushort[] Indices = indices;
 }
